pyproject.toml
[tool.poetry]
name = "fake-service"
version = "0.1.0"
description = ""
authors = ["Jefferson Fragoso de Souza <jefferson.fragoso.ext@jeitto.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.10.6"
fastapi = "^0.115.7"
pydantic-settings = "^2.7.1"
uvicorn = "^0.34.0"
uvloop = "^0.21.0"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
sqlmodel = "^0.0.22"
passlib = "^1.7.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^6.0.0"

[tool.pytest.ini_options]
pythonpath = "src"
python_files = ["test_*.py", "tests_*.py"]

[tool.black]
line-length = 95

[tool.isort]
profile = "black"
line_length = 95

[tool.flake8]
max-line-length = 95

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
